Final try at fixing md8_4
Confirmed md8_4 is the problem field, now to see if it's just angry at the NPC flag because everything else is done on other fields (even though l swear l've seen it on other fields).

Otherwise, a couple lighting fixes, experimenting with an item change, and undoing the new game level nerf since HP is the only thing that changes that early anyway.